Added interpolation to m2s ini parser
Now there is a flag to enable or disable interpolation, disabled by default.
If enabled, ${foo} will be replaced by value of variable 'foo' in the same section.
If there is not such variable, it will be replaced by value of variable 'foo' in 'DEFAULT' section.
If neither exists, it is keept untouched.