Changes in ctx interval reporting paradigm
Centralized interval reporting.
Epoch length setted in commandline. For now, only ctx interval reports use it, but all the other reports will use it in the future.
Report directory and report directory hierarchy.
<report_dir> This dir name is supplied in commandline. It defaults to PWD.
|--interval_reports
|  |--ctx
|     |--ctx100.interval.report
|     |--ctx101.interval.report
|     |--etc.
|--global_reports
This fixes the problem with interval reporting stats for ctx when using checkpoints.

Bugfix in interval reporting after a reset of stats.
Minor fixes in interval reporting.